refactor: 상세페이지 데이터에 맞게 삭제 코드 수정 (#43)
